Day-2/pyproject.toml
[project]
name = "day-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.99.5",
    "python-dotenv>=1.1.1",
]
